Partial Implementation yet again
