+++ genshin/client/__init__.py
@@ -1,0 +1,5 @@
+from .adapter import *
+from .base import *
+from .cache import *
+from .genshin import *
+from .honkai import *
