+++ genshin/client/genshin/__init__.py
@@ -1,1 +1,2 @@
-from .client import *
+from .chinese import *
+from .overseas import *
