+++ genshin/client/genshin/chinese.py
@@ -1,0 +1,2 @@
+from genshin.client import base
+from genshin.client.genshin import base as genshin_base
